[박준호] 과제 목록, 과제 제출
학생그룹 과제 목록조회및 과제 제출기능 구현

+++ lms/src/main/java/gd/fintech/lms/vo/ReportSubmitFile.java
@@ -7,7 +7,7 @@ public class ReportSubmitFile {
 	private String reportSubmitFileUuid;
 	private String reportSubmitFileOriginal;
 	private int reportSubmitNo;
-	private int reportSubmitFileSize;
+	private long reportSubmitFileSize;
 	private String reportSubmitFileType;
 	private int reportSubmitFileCount;
 	private String reportSubmitFileCreatedate;
